Further fix to the calculation of chunking dimensions #86.
The chunking dimensions calculation was not satisfactory in its previous
form.  Using the total number of dimensions didn't work when using extra
virtual dimensions, and my previous fix of hardcoding to 3 dims didn't
work when the writer was used in single mode (with only 2 dimensions of
array data).  The value used is now calculated using a combination of the
incoming array dimensions plus the mode of the writer.