Adding travis configuration
Travis does not have szip installed as it is not in the ubuntu repos. So only
link against sz if SZIP has been defined in configure.
Travis/ubuntu 12.04 supply hdf5 version 1.8.4 - but built with v1.6 default
API. So for the H5Gopen function we need to specify using API version 2 (1.8)